src/lvmnps/actor/commands/__init__.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
#
# @Author: José Sánchez-Gallego (gallegoj@uw.edu)
# @Date: 2023-11-22
# @Filename: __init__.py
# @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)

from __future__ import annotations

from clu.parsers.click import command_parser as lvmnps_command_parser

from .onoff import cycle, off, on
from .refresh import refresh
from .scripts import scripts
from .status import status
