pyproject.toml
[tool.poetry]
name = "sdss-lvmnps"
version = "1.0.0a0"
description = "A library and actor to communicate with an SDSS-V LVM network power switch"
authors = ["Florian Briegel <briegel@mpia.de>", "José Sánchez-Gallego <gallegoj@uw.edu>", "Changgon Kim <changgonkim@khu.ac.kr>", "Mingyeong Yang <mingyeong@khu.ac.kr>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmnps"
repository = "https://github.com/sdss/lvmnps"
documentation = "https://lvmnps.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmnps", from = "src" }
]
include = ["src/lvmnps/actor/schema.json"]

[tool.poetry.scripts]
lvmnps = "lvmnps.__main__:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
sdsstools = "^1.0.0"
click-default-group = "^1.2.2"
sdss-clu = "^2.0.0"
httpx = ">=0.18.1"
pydantic = "^2.5.2"

[tool.poetry.group.dev.dependencies]
ipython = ">=7.11.0"
ipdb = ">=0.12.3"
black = ">=21.7b0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
coverage = {version = ">=5.0", extras = ["toml"]}
Sphinx = ">=7.0.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = ">=2021.6.18"
nox = ">=2021.6.12"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.3.3"
sphinx-click = ">=3.0.1"
ruff = ">=0.1.0"
sphinx-autodoc-typehints = ">=1.25.2"
invoke = ">=2.2.0"
autodoc-pydantic = "^2.0.1"
pytest-httpx = "^0.27.0"

[tool.black]
line-length = 88
target-version = ['py312']
fast = true

[tool.ruff]
line-length = 88
target-version = 'py312'
select = ["E", "F", "I"]
unfixable = ["F841"]
exclude = ["typings/"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]

[tool.ruff.isort]
known-first-party = ["lvmnps"]
lines-after-imports = 2
section-order = ["future", "standard-library", "typing", "third-party", "sdss", "first-party", "local-folder"]

[tool.ruff.isort.sections]
typing = ["typing"]
sdss = ["sdsstools", "clu"]

[tool.pytest.ini_options]
addopts = "--cov lvmnps --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
disable_warnings = ["include-ignored"]
omit = [
    "src/lvmnps/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING",
    "raise NotImplementedError()",
    "pass"
]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
